requirements.txt
ruamel.yaml~=0.17.21
ruamel.yaml.clib~=0.2.7
aiohttp~=3.8.3
tenacity~=8.1.0
yarl~=1.8.2
python-dateutil~=2.8.2
setuptools~=65.5.1